organizando código
